requirements.txt
gradio
openai
pypdf
python-dotenv
pydantic
